# Add doc-values aggregators for the arbitrary aggregation.