## Summary of the changes / Why this improves CrateDB

It improves the performance of the global standard deviation aggregations.
See https://github.com/crate/crate/pull/10048 for more information.

Without doc-values aggregators:
```
# Running setUp
# Running benchmark

## Running Query:
   Name: global arbitrary on long
   Statement:
     select arbitrary(duration) from uservisits
   Concurrency: 1
   Iterations: 2000
100%| 2000/2000 [01:31<00:00, 21.86 requests/s]
Runtime (in ms):
    mean:    44.726 ± 1.202
    min/max: 34.335 → 772.710
Percentile:
    50:   40.981 ± 27.433 (stdev)
    95:   56.560
    99.9: 369.586

## Running Query:
   Name: global arbitrary on float
   Statement:
     select arbitrary("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 2000
100%| 2000/2000 [02:35<00:00, 12.86 requests/s]
Runtime (in ms):
    mean:    74.985 ± 1.459
    min/max: 41.575 → 323.320
Percentile:
    50:   66.986 ± 33.282 (stdev)
    95:   137.141
    99.9: 321.631
# Running tearDown
```

With doc-values aggregators:
```
# Running setUp
# Running benchmark

## Running Query:
   Name: global arbitrary on long
   Statement:
     select arbitrary(duration) from uservisits
   Concurrency: 1
   Iterations: 2000
100%| 2000/2000 [00:27<00:00, 72.97 requests/s]
Runtime (in ms):
    mean:    10.257 ± 0.466
    min/max: 4.806 → 195.278
Percentile:
    50:   8.069 ± 10.635 (stdev)
    95:   18.638
    99.9: 133.295

## Running Query:
   Name: global arbitrary on float
   Statement:
     select arbitrary("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 2000
100%| 2000/2000 [00:35<00:00, 56.06 requests/s]
Runtime (in ms):
    mean:    14.503 ± 0.788
    min/max: 6.206 → 438.253
Percentile:
    50:   12.200 ± 17.986 (stdev)
    95:   22.720
    99.9: 167.751
```
## Checklist

 - [ ] Added an entry in `CHANGES.txt` for user facing changes
 - [ ] Updated documentation & `sql_features` table for user facing changes
 - [x] Touched code is covered by tests
 - [x] [CLA](https://crate.io/community/contribute/cla/) is signed
 - [x] This does not contain breaking changes, or if it does:
    - It is released within a major release
    - It is recorded in ``CHANGES.txt``
    - It was marked as deprecated in an earlier release if possible
    - You've thought about the consequences and other components are adapted
      (E.g. AdminUI)