@kovrus 2020-08-18
retest this please

`org.elasticsearch.index.engine.InternalEngineTests.testLookupSeqNoByIdInLucene` fails with NPE